add uploading photo to server and missing login when token is up-to-date
