+++ app/src/main/java/com/softdesign/devintensive/data/managers/DataManager.java
@@ -5,9 +5,12 @@ import android.content.Context;
 import com.softdesign.devintensive.data.network.RestService;
 import com.softdesign.devintensive.data.network.ServiceGenerator;
 import com.softdesign.devintensive.data.network.requests.UserLoginReq;
+import com.softdesign.devintensive.data.network.responses.UploadPhotoRes;
 import com.softdesign.devintensive.data.network.responses.UserModelRes;
+import com.softdesign.devintensive.data.network.responses.UserRes;
 import com.softdesign.devintensive.utils.DevintensiveApplication;
 
+import okhttp3.MultipartBody;
 import retrofit2.Call;
 
 public class DataManager {
@@ -38,10 +41,26 @@ public class DataManager {
         return mContext;
     }
 
+    public RestService getRestService() {
+        return mRestService;
+    }
+
     //region ==================== Network =======================
 
     public Call<UserModelRes> loginUser(UserLoginReq userLoginReq){
         return mRestService.loginUser(userLoginReq);
+    }
+
+    public Call<UserRes> loginToken(String userId){
+        return mRestService.loginToken(userId);
+    }
+
+    public Call<UploadPhotoRes> uploadPhoto(String userId, MultipartBody.Part file){
+        return mRestService.uploadPhoto(userId, file);
+    }
+
+    public Call<UploadPhotoRes> uploadAvatar(String userId, MultipartBody.Part file){
+        return mRestService.uploadAvatar(userId, file);
     }
 
     //end region
