+++ app/src/main/java/com/softdesign/devintensive/data/network/RestService.java
@@ -1,16 +1,34 @@
 package com.softdesign.devintensive.data.network;
 
+import okhttp3.MultipartBody;
 import retrofit2.Call;
 import retrofit2.http.Body;
-import retrofit2.http.Header;
-import retrofit2.http.Headers;
+import retrofit2.http.GET;
+import retrofit2.http.Multipart;
 import retrofit2.http.POST;
+import retrofit2.http.Part;
+import retrofit2.http.Path;
 
 import com.softdesign.devintensive.data.network.requests.UserLoginReq;
+import com.softdesign.devintensive.data.network.responses.UploadPhotoRes;
 import com.softdesign.devintensive.data.network.responses.UserModelRes;
+import com.softdesign.devintensive.data.network.responses.UserRes;
 
 public interface RestService {
 
     @POST("login")
-    Call<UserModelRes> loginUser (@Body UserLoginReq req);
+    Call<UserModelRes> loginUser(@Body UserLoginReq req);
+
+    @GET("user/{userId}")
+    Call<UserRes> loginToken(@Path("userId") String userId);
+
+    @Multipart
+    @POST("user/{userId}/publicValues/profilePhoto")
+    Call<UploadPhotoRes> uploadPhoto(@Path("userId") String userId,
+                                     @Part MultipartBody.Part file);
+
+    @Multipart
+    @POST("user/{userId}/publicValues/profileAvatar")
+    Call<UploadPhotoRes> uploadAvatar(@Path("userId") String userId,
+                                     @Part MultipartBody.Part file);
 }
