+++ app/src/main/java/com/softdesign/devintensive/data/network/responses/UploadPhotoRes.java
@@ -1,0 +1,22 @@
+package com.softdesign.devintensive.data.network.responses;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+public class UploadPhotoRes {
+    @SerializedName("success")
+    @Expose
+    private boolean success;
+    @SerializedName("data")
+    @Expose
+    private Data data;
+
+    public class Data {
+        @SerializedName("photo")
+        @Expose
+        private String photo;
+        @SerializedName("updated")
+        @Expose
+        private String updated;
+    }
+}
