+++ app/src/main/java/com/softdesign/devintensive/ui/activities/MainActivity.java
@@ -6,6 +6,7 @@ import android.content.ContentValues;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.pm.PackageManager;
+import android.database.Cursor;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.graphics.Color;
@@ -40,6 +41,7 @@ import android.widget.TextView;
 
 import com.softdesign.devintensive.R;
 import com.softdesign.devintensive.data.managers.DataManager;
+import com.softdesign.devintensive.data.network.responses.UploadPhotoRes;
 import com.softdesign.devintensive.ui.custom.EditTextWatcher;
 import com.softdesign.devintensive.ui.custom.RoundedDrawable;
 import com.softdesign.devintensive.utils.ConstantManager;
@@ -55,6 +57,13 @@ import java.util.List;
 import butterknife.BindView;
 import butterknife.BindViews;
 import butterknife.ButterKnife;
+import okhttp3.MediaType;
+import okhttp3.MultipartBody;
+import okhttp3.RequestBody;
+import okhttp3.ResponseBody;
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
 
 public class MainActivity extends BaseActivity implements View.OnClickListener {
     private static final String TAG = ConstantManager.TAG_PREFIX + "MainActivity";
@@ -127,6 +136,7 @@ public class MainActivity extends BaseActivity implements View.OnClickListener {
         Picasso.with(this)
                 .load(mDataManager.getPreferencesManager().loadUserPhoto())
                 .into(mProfileImage);
+        mCollapsingToolbar.setTitle(mDataManager.getPreferencesManager().getUserName());
 
         if (savedInstanceState == null) {
             //активити запускается впервые
@@ -256,12 +266,14 @@ public class MainActivity extends BaseActivity implements View.OnClickListener {
                 if (resultCode == RESULT_OK && data != null){
                     mSelectedImage = data.getData();
                     insertProfileImage(mSelectedImage);
+                    uploadPhoto(getFileFromUri(mSelectedImage));
                 }
                 break;
             case ConstantManager.REQUEST_CAMERA_PICTURE:
                 if (resultCode == RESULT_OK && mPhotoFile != null){
                     mSelectedImage = Uri.fromFile(mPhotoFile);
                     insertProfileImage(mSelectedImage);
+                    uploadPhoto(mPhotoFile);
                 }
                 break;
         }
@@ -440,6 +452,52 @@ public class MainActivity extends BaseActivity implements View.OnClickListener {
                         }
                     }).show();
         }
+    }
+
+    private void uploadPhoto(File file){
+        RequestBody requestFile = RequestBody.create(MediaType.parse("multipart/form-data"), file);
+        MultipartBody.Part body = MultipartBody.Part.createFormData("photo", file.getName(), requestFile);
+        Call<UploadPhotoRes> call = mDataManager.getRestService().uploadPhoto(
+                mDataManager.getPreferencesManager().getUserId(), body);
+        call.enqueue(new Callback<UploadPhotoRes>() {
+            @Override
+            public void onResponse(Call<UploadPhotoRes> call, Response<UploadPhotoRes> response) {
+                //// TODO: 13.07.2016 реализовать получение колбека
+            }
+
+            @Override
+            public void onFailure(Call<UploadPhotoRes> call, Throwable t) {
+                //// TODO: 13.07.2016 реализовать получение ошибки
+            }
+        });
+    }
+
+    private void uploadAvatar(File file){
+        RequestBody requestFile = RequestBody.create(MediaType.parse("multipart/form-data"), file);
+        MultipartBody.Part body = MultipartBody.Part.createFormData("avatar", file.getName(), requestFile);
+        Call<UploadPhotoRes> call = mDataManager.getRestService().uploadPhoto(
+                mDataManager.getPreferencesManager().getUserId(), body);
+        call.enqueue(new Callback<UploadPhotoRes>() {
+            @Override
+            public void onResponse(Call<UploadPhotoRes> call, Response<UploadPhotoRes> response) {
+                //// TODO: 13.07.2016 реализовать получение колбека
+            }
+
+            @Override
+            public void onFailure(Call<UploadPhotoRes> call, Throwable t) {
+                //// TODO: 13.07.2016 реализовать получение ошибки
+            }
+        });
+    }
+
+    private File getFileFromUri(Uri uri){
+        String[] filePathColumn = { MediaStore.Images.Media.DATA };
+        Cursor cursor = this.getContentResolver().query(uri, filePathColumn, null, null, null);
+        cursor.moveToFirst();
+        int columnIndex = cursor.getColumnIndex(filePathColumn[0]);
+        String filePath = cursor.getString(columnIndex);
+        cursor.close();
+        return new File(filePath);
     }
 
     @Override
